requirements.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib
langdetect
googletrans==4.0.0-rc1
